requirements.txt
streamlit
google-generativeai
python-dotenv
Pillow
opencv-python
# SpeechRecognition
pydub
pytesseract